requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
build==0.8.0
    # via pip-tools
click==8.1.3
    # via pip-tools
colorama==0.4.5
    # via
    #   build
    #   click
django==3.2.16
    # via
    #   -r requirements.in
    #   django-js-asset
django-autoslug==1.9.8
    # via -r requirements.in
django-js-asset==2.0.0
    # via django-mptt
django-mptt==0.13.4
    # via -r requirements.in
environs==9.4.0
    # via -r requirements.in
faker==15.1.1
    # via -r requirements.in
marshmallow==3.18.0
    # via environs
mysqlclient==2.1.0
    # via -r requirements.in
packaging==21.3
    # via
    #   build
    #   marshmallow
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via faker
python-dotenv==0.21.0
    # via environs
pytz==2022.4
    # via django
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via
    #   build
    #   pep517
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
